+++ apps/decodex-publisher/src/social_validation/post/source_refs.rs
@@ -1,0 +1,42 @@
+use crate::social_validation::{self, Value};
+
+pub(super) fn validate_social_post_source_refs(refs: Option<&Value>, errors: &mut Vec<String>) {
+	let Some(refs) = refs.and_then(Value::as_object) else {
+		errors.push("source_refs must be an object".into());
+
+		return;
+	};
+	let has_refs = [
+		"reservations",
+		"signals",
+		"social_candidates",
+		"upstream_impacts",
+		"upstream_reviews",
+		"urls",
+	]
+	.iter()
+	.any(|field| {
+		refs.get(*field)
+			.is_some_and(|value| !social_validation::is_empty_or_missing_array(Some(value)))
+	});
+
+	if !has_refs {
+		errors.push(
+			"source_refs must include reservations, signals, social_candidates, upstream_impacts, upstream_reviews, or urls"
+				.into(),
+		);
+	}
+	if refs.get("urls").is_some_and(|urls| !social_validation::is_https_string_array(urls)) {
+		errors.push("source_refs.urls must be a list of https URLs".into());
+	}
+
+	for field in
+		["reservations", "signals", "social_candidates", "upstream_impacts", "upstream_reviews"]
+	{
+		social_validation::validate_optional_string_list(
+			refs.get(field),
+			&format!("source_refs.{field}"),
+			errors,
+		);
+	}
+}
