+++ apps/decodex-publisher/src/social_validation/post/text.rs
@@ -1,0 +1,48 @@
+use crate::social_validation::{self, Value};
+
+pub(in crate::social_validation) fn validate_social_post_text(
+	text: Option<&Value>,
+	errors: &mut Vec<String>,
+) {
+	let Some(items) = social_validation::non_empty_array(text) else {
+		errors.push("text must be a non-empty list of X-sized strings".into());
+
+		return;
+	};
+
+	for (index, text) in items.iter().enumerate() {
+		let Some(text) = text.as_str() else {
+			errors.push(format!("text[{index}] must be a string"));
+
+			continue;
+		};
+
+		validate_social_post_text_item(text, index, errors);
+	}
+}
+
+fn validate_social_post_text_item(text: &str, index: usize, errors: &mut Vec<String>) {
+	if text.is_empty() || text.chars().count() > 280 {
+		errors.push(format!("text[{index}] must be a non-empty X-sized string"));
+	}
+	if text.contains("Automated by @hackink") {
+		errors.push(format!("text[{index}] must not include automation attribution"));
+	}
+	if text.chars().count() > 260 && !text.contains("https://") {
+		errors.push(format!(
+			"text[{index}] longer than 260 characters must include an unavoidable direct source URL"
+		));
+	}
+
+	let normalized = text.trim().to_ascii_lowercase();
+
+	if normalized == "watching this"
+		|| normalized.starts_with("watching this.")
+		|| normalized.starts_with("tracking this.")
+		|| normalized.contains("new release available")
+	{
+		errors.push(format!(
+			"text[{index}] must name a concrete source-backed release, PR, protocol surface, workflow impact, or operator action"
+		));
+	}
+}
